refactor(claude): consolidate batch/client files and embed createBatchMessage example output
- fold batch_client.go and create_batch_message_poll.go into
  client.go / create_batch_message.go respectively (and their tests),
  reducing the package's file count
- move createBatchMessage's ExampleOutput to the embedded-JSON
  pattern already used by textPrompt, instead of an inline struct literal
