+++ docs/components/Claude.mdx
@@ -75,21 +75,20 @@ Emits a single payload once the batch ends, containing:
 ```json
 {
   "data": {
-    "status": "ended",
     "batchId": "msgbatch_01HkcTjaV5uDC8jWR4ZsqFqz",
     "requestCounts": {
-      "processing": 0,
-      "succeeded": 2,
-      "errored": 0,
       "canceled": 0,
-      "expired": 0
+      "errored": 0,
+      "expired": 0,
+      "processing": 0,
+      "succeeded": 2
     },
     "results": [
       {
         "index": 0,
-        "type": "succeeded",
-        "text": "Paris is the capital of France.",
         "stopReason": "end_turn",
+        "text": "Paris is the capital of France.",
+        "type": "succeeded",
         "usage": {
           "input_tokens": 12,
           "output_tokens": 9
@@ -97,15 +96,16 @@ Emits a single payload once the batch ends, containing:
       },
       {
         "index": 1,
-        "type": "succeeded",
-        "text": "Berlin is the capital of Germany.",
         "stopReason": "end_turn",
+        "text": "Berlin is the capital of Germany.",
+        "type": "succeeded",
         "usage": {
           "input_tokens": 12,
           "output_tokens": 9
         }
       }
-    ]
+    ],
+    "status": "ended"
   },
   "timestamp": "2026-07-07T12:00:00Z",
   "type": "claude.createBatchMessage.result"
